Add tests for chat controller
